ocr/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.3
filelock==3.14.0
filetype==1.2.0
fsspec==2024.3.1
ftfy==6.2.0
h11==0.14.0
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
surya-ocr==0.4.0
sympy==1.12
tabulate==0.9.0
tokenizers==0.15.2
torch==2.3.0
tqdm==4.66.2
transformers==4.36.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
